:hamster: Add Pathfinder class.
Do you know pathfinder?
He is one of the Legends in the Apex Legends.
He finds a way/path.
This class also finds paths securely.
The paths found are always under sub directory of the application.
This means this class never find a path which are not related with this application.